requirements.txt
matplotlib~=3.5.1
numpy~=1.21.6
pandas~=1.3.5
scikit-learn~=1.0.2
tensorflow~=2.8.0
scipy~=1.7.3
click~=8.0.3
prettytable~=3.0.0
sacred~=0.8.2
sklearn-evaluation~=0.5.7
dill~=0.3.5.1
ConfigSpace~=0.5.0
